Phase 1: Architecture Decomposition – header split and module map
Implements Phase 1 of the modernization roadmap (tasks 1.1, 1.2, 1.3 and
roadmap ticket 5).

New files
---------
bwgame_tables.h
  Extracts the seven static offline-computed lookup tables from bwgame.h
  (tan_table, direction_table, repulse_adjust_table, cardinal_direction_xy,
  hit_near_target_positions, psi_field_mask, all_player_slots).  Each table
  is annotated with its provenance and intended invariants (task 1.3 –
  data table and constant isolation).

bwgame_state.h
  Extracts all state type definitions from bwgame.h (global_state,
  game_state, state_base_copyable, state_base_non_copyable, state, and
  the autocast helper).  Adds an ownership-model doc-comment that
  formally documents the const/mutable contract for each state layer
  (task 1.2 – explicit ownership and lifetimes).

sync_protocol.h
  Extracts the sync_messages enum namespace from sync.h.  Adds a
  wire-format stability contract (no reordering, append-only additions)
  and per-constant comments (roadmap ticket 4 / Phase 2 groundwork).

docs/architecture.md
  New module map document describing every subsystem, its header(s),
  the ownership/lifetime model, the simplified include-order graph, and
  a summary of Phase 1 changes (roadmap ticket 5).

Changed files
-------------
bwgame.h
  Converted to a compatibility facade: now opens with includes of
  bwgame_tables.h and bwgame_state.h; the extracted definitions are
  replaced with redirect comments.  All existing consumers that include
  "bwgame.h" compile without modification (task 1.1 – compatibility
  facade).

sync.h
  Adds #include "sync_protocol.h"; replaces the inline sync_messages
  enum block with a redirect comment.

No simulation logic was changed; deterministic behaviour is preserved.
