+++ bwgame.h
@@ -1,11 +1,24 @@
 #ifndef BWGAME_BWGAME_H
 #define BWGAME_BWGAME_H
 
-#include "util.h"
-#include "data_types.h"
-#include "game_types.h"
+// bwgame.h – compatibility facade.
+//
+// The simulation engine was historically a single monolithic header.  Phase 1
+// of the modernization roadmap decomposed it into focused modules:
+//
+//   bwgame_tables.h  – offline-computed lookup tables (angles, directions,
+//                      combat scatter offsets, psi-field masks, player slots).
+//   bwgame_state.h   – state type definitions with documented ownership model
+//                      (global_state, game_state, state_base_copyable,
+//                       state_base_non_copyable, state).
+//
+// All downstream consumers that include "bwgame.h" continue to work without
+// change.  New code should prefer including the narrower headers directly.
+
+#include "bwgame_tables.h"
+#include "bwgame_state.h"
+
 #include "data_loading.h"
-#include "bwenums.h"
 #include "korean.h"
 
 #include <algorithm>
@@ -16,318 +29,9 @@
 
 namespace bwgame {
 
-static const std::array<unsigned int, 64> tan_table = {
-	7, 13, 19, 26, 32, 38, 45, 51, 58, 65, 71, 78, 85, 92,
-	99, 107, 114, 122, 129, 137, 146, 154, 163, 172, 181,
-	190, 200, 211, 221, 233, 244, 256, 269, 283, 297, 312,
-	329, 346, 364, 384, 405, 428, 452, 479, 509, 542, 578,
-	619, 664, 716, 775, 844, 926, 1023, 1141, 1287, 1476,
-	1726, 2076, 2600, 3471, 5211, 10429, std::numeric_limits<unsigned int>::max()
-};
+// (lookup tables relocated to bwgame_tables.h – included above via the facade)
 
-static const std::array<int, 16 * 2> repulse_adjust_table = {
-	-5, -5, -5, 5, 5, -5, 5, 5, 5, -5, 5, 5, -5, -5, -5, 5, -5, 5, 5, -5, -5, -5, 5, 5, -5, 5, 5, -5, 5, 5, -5, -5
-};
-
-static const std::array<xy, 4> cardinal_direction_xy = {xy{1, 0}, xy{0, 1}, xy{-1, 0}, xy{0, -1}};
-
-static const xy_fp8 direction_table[256] = {
-	{0_fp8,-256_fp8},{6_fp8,-256_fp8},{13_fp8,-256_fp8},{19_fp8,-255_fp8},{25_fp8,-255_fp8},{31_fp8,-254_fp8},{38_fp8,-253_fp8},{44_fp8,-252_fp8},
-	{50_fp8,-251_fp8},{56_fp8,-250_fp8},{62_fp8,-248_fp8},{68_fp8,-247_fp8},{74_fp8,-245_fp8},{80_fp8,-243_fp8},{86_fp8,-241_fp8},{92_fp8,-239_fp8},
-	{98_fp8,-237_fp8},{104_fp8,-234_fp8},{109_fp8,-231_fp8},{115_fp8,-229_fp8},{121_fp8,-226_fp8},{126_fp8,-223_fp8},{132_fp8,-220_fp8},{137_fp8,-216_fp8},
-	{142_fp8,-213_fp8},{147_fp8,-209_fp8},{152_fp8,-206_fp8},{157_fp8,-202_fp8},{162_fp8,-198_fp8},{167_fp8,-194_fp8},{172_fp8,-190_fp8},{177_fp8,-185_fp8},
-	{181_fp8,-181_fp8},{185_fp8,-177_fp8},{190_fp8,-172_fp8},{194_fp8,-167_fp8},{198_fp8,-162_fp8},{202_fp8,-157_fp8},{206_fp8,-152_fp8},{209_fp8,-147_fp8},
-	{213_fp8,-142_fp8},{216_fp8,-137_fp8},{220_fp8,-132_fp8},{223_fp8,-126_fp8},{226_fp8,-121_fp8},{229_fp8,-115_fp8},{231_fp8,-109_fp8},{234_fp8,-104_fp8},
-	{237_fp8,-98_fp8},{239_fp8,-92_fp8},{241_fp8,-86_fp8},{243_fp8,-80_fp8},{245_fp8,-74_fp8},{247_fp8,-68_fp8},{248_fp8,-62_fp8},{250_fp8,-56_fp8},
-	{251_fp8,-50_fp8},{252_fp8,-44_fp8},{253_fp8,-38_fp8},{254_fp8,-31_fp8},{255_fp8,-25_fp8},{255_fp8,-19_fp8},{256_fp8,-13_fp8},{256_fp8,-6_fp8},
-	{256_fp8,0_fp8},{256_fp8,6_fp8},{256_fp8,13_fp8},{255_fp8,19_fp8},{255_fp8,25_fp8},{254_fp8,31_fp8},{253_fp8,38_fp8},{252_fp8,44_fp8},
-	{251_fp8,50_fp8},{250_fp8,56_fp8},{248_fp8,62_fp8},{247_fp8,68_fp8},{245_fp8,74_fp8},{243_fp8,80_fp8},{241_fp8,86_fp8},{239_fp8,92_fp8},
-	{237_fp8,98_fp8},{234_fp8,104_fp8},{231_fp8,109_fp8},{229_fp8,115_fp8},{226_fp8,121_fp8},{223_fp8,126_fp8},{220_fp8,132_fp8},{216_fp8,137_fp8},
-	{213_fp8,142_fp8},{209_fp8,147_fp8},{206_fp8,152_fp8},{202_fp8,157_fp8},{198_fp8,162_fp8},{194_fp8,167_fp8},{190_fp8,172_fp8},{185_fp8,177_fp8},
-	{181_fp8,181_fp8},{177_fp8,185_fp8},{172_fp8,190_fp8},{167_fp8,194_fp8},{162_fp8,198_fp8},{157_fp8,202_fp8},{152_fp8,206_fp8},{147_fp8,209_fp8},
-	{142_fp8,213_fp8},{137_fp8,216_fp8},{132_fp8,220_fp8},{126_fp8,223_fp8},{121_fp8,226_fp8},{115_fp8,229_fp8},{109_fp8,231_fp8},{104_fp8,234_fp8},
-	{98_fp8,237_fp8},{92_fp8,239_fp8},{86_fp8,241_fp8},{80_fp8,243_fp8},{74_fp8,245_fp8},{68_fp8,247_fp8},{62_fp8,248_fp8},{56_fp8,250_fp8},
-	{50_fp8,251_fp8},{44_fp8,252_fp8},{38_fp8,253_fp8},{31_fp8,254_fp8},{25_fp8,255_fp8},{19_fp8,255_fp8},{13_fp8,256_fp8},{6_fp8,256_fp8},
-	{0_fp8,256_fp8},{-6_fp8,256_fp8},{-13_fp8,256_fp8},{-19_fp8,255_fp8},{-25_fp8,255_fp8},{-31_fp8,254_fp8},{-38_fp8,253_fp8},{-44_fp8,252_fp8},
-	{-50_fp8,251_fp8},{-56_fp8,250_fp8},{-62_fp8,248_fp8},{-68_fp8,247_fp8},{-74_fp8,245_fp8},{-80_fp8,243_fp8},{-86_fp8,241_fp8},{-92_fp8,239_fp8},
-	{-98_fp8,237_fp8},{-104_fp8,234_fp8},{-109_fp8,231_fp8},{-115_fp8,229_fp8},{-121_fp8,226_fp8},{-126_fp8,223_fp8},{-132_fp8,220_fp8},{-137_fp8,216_fp8},
-	{-142_fp8,213_fp8},{-147_fp8,209_fp8},{-152_fp8,206_fp8},{-157_fp8,202_fp8},{-162_fp8,198_fp8},{-167_fp8,194_fp8},{-172_fp8,190_fp8},{-177_fp8,185_fp8},
-	{-181_fp8,181_fp8},{-185_fp8,177_fp8},{-190_fp8,172_fp8},{-194_fp8,167_fp8},{-198_fp8,162_fp8},{-202_fp8,157_fp8},{-206_fp8,152_fp8},{-209_fp8,147_fp8},
-	{-213_fp8,142_fp8},{-216_fp8,137_fp8},{-220_fp8,132_fp8},{-223_fp8,126_fp8},{-226_fp8,121_fp8},{-229_fp8,115_fp8},{-231_fp8,109_fp8},{-234_fp8,104_fp8},
-	{-237_fp8,98_fp8},{-239_fp8,92_fp8},{-241_fp8,86_fp8},{-243_fp8,80_fp8},{-245_fp8,74_fp8},{-247_fp8,68_fp8},{-248_fp8,62_fp8},{-250_fp8,56_fp8},
-	{-251_fp8,50_fp8},{-252_fp8,44_fp8},{-253_fp8,38_fp8},{-254_fp8,31_fp8},{-255_fp8,25_fp8},{-255_fp8,19_fp8},{-256_fp8,13_fp8},{-256_fp8,6_fp8},
-	{-256_fp8,0_fp8},{-256_fp8,-6_fp8},{-256_fp8,-13_fp8},{-255_fp8,-19_fp8},{-255_fp8,-25_fp8},{-254_fp8,-31_fp8},{-253_fp8,-38_fp8},{-252_fp8,-44_fp8},
-	{-251_fp8,-50_fp8},{-250_fp8,-56_fp8},{-248_fp8,-62_fp8},{-247_fp8,-68_fp8},{-245_fp8,-74_fp8},{-243_fp8,-80_fp8},{-241_fp8,-86_fp8},{-239_fp8,-92_fp8},
-	{-237_fp8,-98_fp8},{-234_fp8,-104_fp8},{-231_fp8,-109_fp8},{-229_fp8,-115_fp8},{-226_fp8,-121_fp8},{-223_fp8,-126_fp8},{-220_fp8,-132_fp8},{-216_fp8,-137_fp8},
-	{-213_fp8,-142_fp8},{-209_fp8,-147_fp8},{-206_fp8,-152_fp8},{-202_fp8,-157_fp8},{-198_fp8,-162_fp8},{-194_fp8,-167_fp8},{-190_fp8,-172_fp8},{-185_fp8,-177_fp8},
-	{-181_fp8,-181_fp8},{-177_fp8,-185_fp8},{-172_fp8,-190_fp8},{-167_fp8,-194_fp8},{-162_fp8,-198_fp8},{-157_fp8,-202_fp8},{-152_fp8,-206_fp8},{-147_fp8,-209_fp8},
-	{-142_fp8,-213_fp8},{-137_fp8,-216_fp8},{-132_fp8,-220_fp8},{-126_fp8,-223_fp8},{-121_fp8,-226_fp8},{-115_fp8,-229_fp8},{-109_fp8,-231_fp8},{-104_fp8,-234_fp8},
-	{-98_fp8,-237_fp8},{-92_fp8,-239_fp8},{-86_fp8,-241_fp8},{-80_fp8,-243_fp8},{-74_fp8,-245_fp8},{-68_fp8,-247_fp8},{-62_fp8,-248_fp8},{-56_fp8,-250_fp8},
-	{-50_fp8,-251_fp8},{-44_fp8,-252_fp8},{-38_fp8,-253_fp8},{-31_fp8,-254_fp8},{-25_fp8,-255_fp8},{-19_fp8,-255_fp8},{-13_fp8,-256_fp8},{-6_fp8,-256_fp8}
-};
-
-static const xy hit_near_target_positions[14] = {
-	{-1, -1}, {-51, -31}, {-7, -12}, {36, 52}, {49, -25}, {-40, 46}, {4, 19}, {-28, -50},
-	{-50, 15}, {54, 13}, {-11, 53}, {-15, 9}, {19, -50}, {17, -8}
-};
-
-static const bool psi_field_mask[5][8] = {
-	{ 1, 1, 1, 1, 1, 1, 1, 1 },
-	{ 1, 1, 1, 1, 1, 1, 1, 1 },
-	{ 1, 1, 1, 1, 1, 1, 1, 0 },
-	{ 1, 1, 1, 1, 1, 1, 0, 0 },
-	{ 1, 1, 1, 0, 0, 0, 0, 0 }
-};
-
-static const std::array<int, 12> all_player_slots = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
-
-template<typename T>
-struct autocast {
-	T val;
-	operator T() {
-		return val;
-	}
-	T operator->() {
-		return val;
-	}
-	autocast(T val) : val(val) {}
-	template<typename T2, typename std::enable_if<std::is_same<typename std::decay<T2>::type, unit_t>::value>::type* = nullptr>
-	autocast(T2* ptr) : val(ptr->unit_type) {}
-	template<typename T2>
-	autocast(type_id<T2> ptr) : val(ptr) {}
-};
-
-using unit_type_autocast = autocast<const unit_type_t*>;
-
-struct global_state {
-
-	global_state() = default;
-	global_state(global_state&) = delete;
-	global_state(global_state&&) = default;
-	global_state& operator=(global_state&) = delete;
-	global_state& operator=(global_state&&) = default;
-
-	flingy_types_t flingy_types;
-	sprite_types_t sprite_types;
-	image_types_t image_types;
-	order_types_t order_types;
-	iscript_t iscript;
-
-	a_vector<grp_t> grps;
-	a_vector<grp_t*> image_grp;
-	a_vector<a_vector<a_vector<xy>>> lo_offsets;
-	a_vector<std::array<a_vector<a_vector<xy>>*, 6>> image_lo_offsets;
-
-	a_vector<uint8_t> units_dat;
-	a_vector<uint8_t> weapons_dat;
-	a_vector<uint8_t> upgrades_dat;
-	a_vector<uint8_t> techdata_dat;
-
-	a_vector<uint8_t> melee_trg;
-
-	std::array<a_vector<uint8_t>, 8> tileset_vf4;
-	std::array<a_vector<uint8_t>, 8> tileset_cv5;
-};
-
-struct game_state {
-
-	game_state() = default;
-	game_state(const game_state&) = delete;
-	game_state(game_state&&) = default;
-	game_state& operator=(const game_state&) = delete;
-	game_state& operator=(game_state&&) = default;
-
-	size_t map_tile_width;
-	size_t map_tile_height;
-	size_t map_walk_width;
-	size_t map_walk_height;
-	size_t map_width;
-	size_t map_height;
-
-	a_vector<a_string> map_strings;
-
-	a_string scenario_name;
-	a_string scenario_description;
-
-	type_indexed_array<int, UnitTypes> unit_air_strength;
-	type_indexed_array<int, UnitTypes> unit_ground_strength;
-
-	struct force_t {
-		a_string name;
-		uint8_t flags;
-	};
-	std::array<force_t, 4> forces;
-
-	std::array<sight_values_t, 12> sight_values;
-
-	size_t tileset_index;
-
-	a_vector<tile_id> gfx_tiles;
-	a_vector<cv5_entry> cv5;
-	a_vector<vf4_entry> vf4;
-	a_vector<uint16_t> mega_tile_flags;
-
-	unit_types_t unit_types;
-	weapon_types_t weapon_types;
-	upgrade_types_t upgrade_types;
-	tech_types_t tech_types;
-
-	std::array<type_indexed_array<bool, UnitTypes>, 12> unit_type_allowed;
-	std::array<type_indexed_array<int, UpgradeTypes>, 12> max_upgrade_levels;
-	std::array<type_indexed_array<bool, TechTypes>, 12> tech_available;
-
-	std::array<xy, 12> start_locations;
-
-	int max_unit_width;
-	int max_unit_height;
-
-	size_t repulse_field_width;
-	size_t repulse_field_height;
-
-	regions_t regions;
-
-	a_vector<trigger> triggers;
-};
-
-struct state_base_copyable {
-
-	const global_state* global;
-	game_state* game;
-
-	int update_tiles_countdown;
-
-	int order_timer_counter;
-	int secondary_order_timer_counter;
-	int current_frame;
-
-	std::array<player_t, 12> players;
-
-	std::array<std::array<int, 12>, 12> alliances;
-
-	std::array<type_indexed_array<int, UpgradeTypes>, 12> upgrade_levels;
-	std::array<type_indexed_array<bool, UpgradeTypes>, 12> upgrade_upgrading;
-	std::array<type_indexed_array<bool, TechTypes>, 12> tech_researched;
-	std::array<type_indexed_array<bool, TechTypes>, 12> tech_researching;
-
-	std::array<type_indexed_array<int, UnitTypes>, 12> unit_counts;
-	std::array<type_indexed_array<int, UnitTypes>, 12> completed_unit_counts;
-
-	std::array<int, 12> factory_counts;
-	std::array<int, 12> building_counts;
-	std::array<int, 12> non_building_counts;
-
-	std::array<int, 12> completed_factory_counts;
-	std::array<int, 12> completed_building_counts;
-	std::array<int, 12> completed_non_building_counts;
-
-	std::array<int, 12> total_buildings_ever_completed;
-	std::array<int, 12> total_non_buildings_ever_completed;
-
-	std::array<int, 12> unit_score;
-	std::array<int, 12> building_score;
-
-	std::array<std::array<fp1, 3>, 12> supply_used;
-	std::array<std::array<fp1, 3>, 12> supply_available;
-
-	std::array<uint32_t, 12> shared_vision;
-
-	a_vector<tile_t> tiles;
-	a_vector<uint16_t> tiles_mega_tile_index;
-
-	std::array<int, 0x100> random_counts;
-	int total_random_counts;
-	uint32_t lcg_rand_state;
-
-	int last_error;
-
-	int trigger_timer;
-	std::array<a_vector<running_trigger>, 8> running_triggers;
-	std::array<int, 12> trigger_wait_timers;
-	std::array<bool, 12> trigger_waiting;
-
-	size_t active_orders_size;
-	size_t active_bullets_size;
-	size_t active_thingies_size;
-
-	a_vector<uint8_t> repulse_field;
-
-	bool prev_bullet_heading_offset_clockwise;
-
-	std::array<int, 12> current_minerals;
-	std::array<int, 12> current_gas;
-	std::array<int, 12> total_minerals_gathered;
-	std::array<int, 12> total_gas_gathered;
-
-	std::array<static_vector<std::pair<size_t, size_t>, 16>, 32> recent_lurker_hits;
-	size_t recent_lurker_hit_current_index;
-
-	creep_life_t creep_life;
-	bool update_psionic_matrix;
-	int disruption_webbed_units;
-	bool cheats_enabled;
-	bool cheat_operation_cwal;
-
-	a_vector<location> locations;
-};
-
-struct psionic_matrix_link_f {
-	auto* operator()(unit_t* ptr) {
-		return &ptr->building.pylon.psionic_matrix_link;
-	}
-	auto* operator()(const unit_t* ptr) {
-		return &ptr->building.pylon.psionic_matrix_link;
-	}
-};
-
-struct state_base_non_copyable {
-
-	state_base_non_copyable() = default;
-	state_base_non_copyable(const state_base_non_copyable&) = delete;
-	state_base_non_copyable(state_base_non_copyable&&) = default;
-	state_base_non_copyable& operator=(const state_base_non_copyable&) = delete;
-	state_base_non_copyable& operator=(state_base_non_copyable&&) = default;
-
-	intrusive_list<unit_t, default_link_f> visible_units;
-	intrusive_list<unit_t, default_link_f> hidden_units;
-	intrusive_list<unit_t, default_link_f> map_revealer_units;
-	intrusive_list<unit_t, default_link_f> dead_units;
-
-	std::array<intrusive_list<unit_t, void, &unit_t::player_units_link>, 12> player_units;
-	intrusive_list<unit_t, void, &unit_t::cloaked_unit_link> cloaked_units;
-	intrusive_list<unit_t, psionic_matrix_link_f> psionic_matrix_units;
-
-	object_container<unit_t, 1700, 17> units_container;
-
-	intrusive_list<bullet_t, default_link_f> active_bullets;
-	object_container<bullet_t, 100, 10> bullets_container;
-
-	a_vector<intrusive_list<sprite_t, default_link_f>> sprites_on_tile_line;
-	object_container<sprite_t, 2500, 25> sprites_container;
-
-	object_container<image_t, 5000, 50> images_container;
-
-	object_container<order_t, 2000, 20> orders_container;
-
-	intrusive_list<path_t, default_link_f> free_paths;
-	a_list<path_t> paths;
-
-	intrusive_list<thingy_t, default_link_f> active_thingies;
-	intrusive_list<thingy_t, default_link_f> free_thingies;
-	a_list<thingy_t> thingies;
-
-	struct unit_finder_entry {
-		unit_t* u;
-		int value;
-	};
-	a_vector<unit_finder_entry> unit_finder_x;
-	a_vector<unit_finder_entry> unit_finder_y;
-
-	const unit_t* consider_collision_with_unit_bug;
-	const unit_t* prev_bullet_source_unit;
-};
-
-struct state : state_base_copyable, state_base_non_copyable {
-};
+// (state type definitions relocated to bwgame_state.h – included above via the facade)
 
 struct state_functions {
